# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Academic Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents for current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/upload:
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document (PDF, DOCX, TXT)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{documentId}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a single document
      security:
        - BearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      security:
        - BearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Summaries ───────────────────────────────────────────────────────────────
...
  /documents/{documentId}/quiz:
    post:
      operationId: generateQuiz
      tags: [quiz]
      summary: Generate a quiz from a document
      security:
        - BearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quiz:
    get:
      operationId: listQuizzes
      tags: [quiz]
      summary: List all quizzes for current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
  /quiz/{quizId}:
    get:
      operationId: getQuiz
      tags: [quiz]
      summary: Get a specific quiz
      security:
        - BearerAuth: []
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quiz/{quizId}/submit:
    post:
      operationId: submitQuiz
      tags: [quiz]
      summary: Submit quiz answers
      security:
        - BearerAuth: []
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"

  # ─── Flashcards ───────────────────────────────────────────────────────────────
  /documents/{documentId}/flashcards:
    post:
      operationId: generateFlashcards
      tags: [flashcards]
      summary: Generate flashcards from a document
      security:
        - BearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
  /flashcards:
    get:
      operationId: listFlashcardSets
      tags: [flashcards]
      summary: List all flashcard sets for current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardSet"
  /flashcards/{setId}:
    get:
      operationId: getFlashcardSet
      tags: [flashcards]
      summary: Get a specific flashcard set
      security:
        - BearerAuth: []
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"

  # ─── Dashboard ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: ["string", "null"]
      required:
        - error

    # ─── Auth ──────────────────────────────────────────────────────────────────
...
    DocumentUploadInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    Document:
      type: object
      required: [id, userId, name, mimeType, size, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        status:
          type: string
          enum: [pending, processing, ready, error]
        pageCount:
          type: ["integer", "null"]
        wordCount:
          type: ["integer", "null"]
        summary:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── Summaries ─────────────────────────────────────────────────────────────
...
    QuizInput:
      type: object
      required: [difficulty, questionCount]
      properties:
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionCount:
          type: integer
          minimum: 1
          maximum: 20
        questionTypes:
          type: array
          items:
            type: string
            enum: [multiple_choice, true_false, short_answer]
    QuizOption:
      type: object
      required: [label, text]
      properties:
        label:
          type: string
        text:
          type: string
    QuizQuestion:
      type: object
      required: [id, question, questionType, answer]
      properties:
        id:
          type: string
        question:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOption"
        answer:
          type: string
        explanation:
          type: ["string", "null"]
    Quiz:
      type: object
      required: [id, documentId, difficulty, questions, createdAt]
      properties:
        id:
          type: string
        documentId:
          type: string
        documentName:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        score:
          type: ["integer", "null"]
        maxScore:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
    QuizResult:
      type: object
      required: [quizId, score, maxScore, percentage, feedback]
      properties:
        quizId:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        percentage:
          type: number
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/QuizFeedbackItem"
    QuizFeedbackItem:
      type: object
      required: [questionId, correct, correctAnswer]
      properties:
        questionId:
          type: string
        correct:
          type: boolean
        yourAnswer:
          type: ["string", "null"]
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]

    # ─── Flashcards ────────────────────────────────────────────────────────────
    FlashcardsInput:
      type: object
      required: [cardCount]
      properties:
        cardCount:
          type: integer
          minimum: 1
          maximum: 50
        topic:
          type: ["string", "null"]
    Flashcard:
      type: object
      required: [id, question, answer]
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
    FlashcardSet:
      type: object
      required: [id, documentId, cards, createdAt]
      properties:
        id:
          type: string
        documentId:
          type: string
        documentName:
          type: ["string", "null"]
        topic:
          type: ["string", "null"]
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        createdAt:
          type: string
          format: date-time

    # ─── Dashboard ─────────────────────────────────────────────────────────────
...