# lib/api-zod/src/generated/types/flashcard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Academic Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Flashcard {
  id: string;
  question: string;
  answer: string;
}