# lib/api-zod/src/generated/types/flashcardSet.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Academic Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from './flashcard';

export interface FlashcardSet {
  id: string;
  documentId: string;
  /** @nullable */
  documentName?: string | null;
  /** @nullable */
  topic?: string | null;
  cards: Flashcard[];
  createdAt: Date;
}