# lib/api-zod/src/generated/types/quizOption.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Academic Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizOption {
  label: string;
  text: string;
}